+++ internal/sdkprovider/service/account/account_test.go
@@ -88,7 +88,7 @@ func testAccCheckAivenAccountResourceDestroy(s *terraform.State) error {
 
 		r, err := c.Accounts.List(ctx)
 		if err != nil {
-			var e *aiven.Error
+			var e aiven.Error
 			if errors.As(err, &e) && e.Status != 404 {
 				return err
 			}