+++ internal/sdkprovider/service/influxdb/influxdb_user_test.go
@@ -57,7 +57,7 @@ func testAccCheckAivenInfluxDBUserResourceDestroy(s *terraform.State) error {
 
 		p, err := c.ServiceUsers.Get(ctx, projectName, serviceName, username)
 		if err != nil {
-			var e *aiven.Error
+			var e aiven.Error
 			if errors.As(err, &e) && e.Status != 404 {
 				return err
 			}