+++ internal/sdkprovider/service/pg/pg_user_test.go
@@ -124,7 +124,7 @@ func testAccCheckAivenPGUserResourceDestroy(s *terraform.State) error {
 
 		p, err := c.ServiceUsers.Get(ctx, projectName, serviceName, username)
 		if err != nil {
-			var e *aiven.Error
+			var e aiven.Error
 			if errors.As(err, &e) && e.Status != 404 {
 				return err
 			}