+++ internal/sdkprovider/service/redis/redis_user_test.go
@@ -56,7 +56,7 @@ func testAccCheckAivenRedisUserResourceDestroy(s *terraform.State) error {
 
 		p, err := c.ServiceUsers.Get(ctx, projectName, serviceName, username)
 		if err != nil {
-			var e *aiven.Error
+			var e aiven.Error
 			if errors.As(err, &e) && e.Status != 404 {
 				return err
 			}