+++ internal/sdkprovider/service/vpc/project_vpc.go
@@ -219,7 +219,7 @@ func (w *ProjectVPCDeleteWaiter) RefreshFunc() resource.StateRefreshFunc {
 
 				// VPC cannot be deleted while there are services migrating from
 				// it or service deletion is still in progress
-				var e *aiven.Error
+				var e aiven.Error
 				if errors.As(err, &e) && e.Status != 409 {
 					return nil, "", err
 				}