+++ CHANGELOG.md
@@ -13,6 +13,7 @@ nav_order: 1
 
 <!-- TODO: uncomment when dragonfly is supported -->
 <!-- - Dragonfly support -->
+- Fix `aiven_kafka_schema` JSON object diff suppress function
 
 ## [4.13.3] - 2024-01-29